requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.10
django-cors-headers==4.6.0
django-rest-swagger==2.2.0
drf-yasg==1.21.8
stripe==11.4.1
daphne==4.1.2
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.121.0  # ✅ Required for Gmail API